first compilation done!

+++ FlowDesigner/FuzzyEngine/src/FuzzySet.cc
@@ -20,7 +20,10 @@
 
 #include "FuzzySet.h"
 #include "Exception.h"
-
+#include <stdlib.h>
+#include <stdio.h>
+#include <string>
+#include <iostream>
 //////////////////////////////////////////////////////////////////////
 // Construction
 //////////////////////////////////////////////////////////////////////
@@ -34,19 +37,21 @@ FuzzySet::FuzzySet(const string &name)
 //////////////////////////////////////////////////////////////////////
 
 FuzzySet::~FuzzySet() {
-
-	for (int i = 0; i < m_functions.size(); i++) {
-		delete m_functions[i];
-	}
-
-	m_functions.resize(0);
+  
+  for (int i = 0; i < m_functions.size(); i++) {
+    delete m_functions[i];
+  }
+  
+  m_functions.resize(0);
 }
 //////////////////////////////////////////////////////////////////////
 // add a trapezoid function
 //////////////////////////////////////////////////////////////////////
 
-void FuzzySet::add_trapezoidal_function(const string &name, float a, float b, float c, float d) {
-	m_functions.push_back(new TrapezoidalFunction(name,a,b,c,d));	
+void FuzzySet::add_trapezoidal_function(const string &name, float a, 
+					float b, float c, float d) {
+
+  m_functions.push_back(new TrapezoidalFunction(name,a,b,c,d));	
 }
 
 //////////////////////////////////////////////////////////////////////
@@ -54,7 +59,7 @@ void FuzzySet::add_trapezoidal_function(const string &name, float a, float b, fl
 //////////////////////////////////////////////////////////////////////
 
 void FuzzySet::add_triangular_function(const string &name, float a, float b, float c) {
-	m_functions.push_back(new TriangularFunction(name,a,b,c));
+  m_functions.push_back(new TriangularFunction(name,a,b,c));
 }
 
 //////////////////////////////////////////////////////////////////////
@@ -62,25 +67,25 @@ void FuzzySet::add_triangular_function(const string &name, float a, float b, flo
 //////////////////////////////////////////////////////////////////////
 
 vector<float> & FuzzySet::get_all_membership_evaluation(float x) {
-
-	//updating evaluation vector
-	m_evaluation.resize(m_functions.size());
-	
-	//updating maps
-	m_string_value_map.clear();
-
-
-
-	for (int i = 0; i < m_functions.size(); i++) {
-		m_evaluation[i] = m_functions[i]->evaluate(x);
-
-		m_string_value_map.insert(
-			pair<string,float>(m_functions[i]->get_name(),m_evaluation[i]));
-
-	}
-	
-	return m_evaluation;
-
+  
+  //updating evaluation vector
+  m_evaluation.resize(m_functions.size());
+  
+  //updating maps
+  m_string_value_map.clear();
+  
+  
+  
+  for (int i = 0; i < m_functions.size(); i++) {
+    m_evaluation[i] = m_functions[i]->evaluate(x);
+    
+    m_string_value_map.insert(
+			      pair<string,float>(m_functions[i]->get_name(),m_evaluation[i]));
+    
+  }
+  
+  return m_evaluation;
+  
 }
 
 //////////////////////////////////////////////////////////////////////
@@ -88,16 +93,16 @@ vector<float> & FuzzySet::get_all_membership_evaluation(float x) {
 //////////////////////////////////////////////////////////////////////
 
 float FuzzySet::get_membership_evaluation(const string &name, float x) {
-
-
-	int index = find_function_by_index(name);
-
-	m_evaluation[index] = m_functions[index]->evaluate(x);
-
-	m_string_value_map.insert(
-			pair<string,float>(name,m_evaluation[index]));
-
-	return m_evaluation[index];
+  
+  
+  int index = find_function_by_index(name);
+  
+  m_evaluation[index] = m_functions[index]->evaluate(x);
+  
+  m_string_value_map.insert(
+			    pair<string,float>(name,m_evaluation[index]));
+  
+  return m_evaluation[index];
 }
 
 
@@ -106,20 +111,20 @@ float FuzzySet::get_membership_evaluation(const string &name, float x) {
 //////////////////////////////////////////////////////////////////////
 
 int FuzzySet::find_function_by_index(const string &name) {
-
-	for (int i = 0; i < m_functions.size(); i++) {
-
-		if (m_functions[i]->get_name() == name) return i;
-
-	}
-
-	char message[256];
-
-	sprintf(message,"CANNOT GET MEMBERSHIP FUNCTION CALLED : %s",name.c_str());
-
-	throw GeneralException(message,__FILE__,__LINE__);
-
-	return -1;
+  
+  for (int i = 0; i < m_functions.size(); i++) {
+    
+    if (m_functions[i]->get_name() == name) return i;
+    
+  }
+  
+  char message[256];
+  
+  sprintf(message,"CANNOT GET MEMBERSHIP FUNCTION CALLED : %s",name.c_str());
+  
+  throw GeneralException(message,__FILE__,__LINE__);
+  
+  return -1;
 }
 
 //////////////////////////////////////////////////////////////////////
@@ -127,67 +132,67 @@ int FuzzySet::find_function_by_index(const string &name) {
 //////////////////////////////////////////////////////////////////////
 
 FuzzyFunction * FuzzySet::find_function_by_name (const string &name) {
-
-	for (int i = 0; i < m_functions.size(); i++) {
-
-		if (m_functions[i]->get_name() == name) return m_functions[i];
-
-	}
-
-	char message[256];
-
-	sprintf(message,"CANNOT GET MEMBERSHIP FUNCTION CALLED : %s",name.c_str());
-
-	throw GeneralException(message,__FILE__,__LINE__);
-
-	return NULL;
-
+  
+  for (int i = 0; i < m_functions.size(); i++) {
+    
+    if (m_functions[i]->get_name() == name) return m_functions[i];
+    
+  }
+  
+  char message[256];
+  
+  sprintf(message,"CANNOT GET MEMBERSHIP FUNCTION CALLED : %s",name.c_str());
+  
+  throw GeneralException(message,__FILE__,__LINE__);
+  
+  return NULL;
+  
 }
 //////////////////////////////////////////////////////////////////////
 // Printing the functions (graphics)
 //////////////////////////////////////////////////////////////////////
 
 void FuzzySet::print_functions(ostream &out) {
-
-	if (!m_functions.empty()) {
-		int min = (int) m_functions[0]->get_lower_bound(); 
-		int max = (int) m_functions[0]->get_upper_bound();
-		int i;
-
-		//finding limits
-		for (i = 0; i < m_functions.size(); i++) {
-
-			if (min > m_functions[i]->get_lower_bound()) {
-				min = m_functions[i]->get_lower_bound();
-			}
-	
-			if (max < m_functions[i]->get_upper_bound()) {
-				max = m_functions[i]->get_upper_bound();
-			}
-
-
-		}
-		
-		//printing membership values
-		for (float index = (float) min; index <= (float) max; index++) {
-		
-			//printing membership function names
-			for (i = 0; i < m_functions.size(); i++) {
-				out<<m_functions[i]->get_name()<<"\t";	
-			}
-
-			out<<(int) index<<"\t";
-
-
-			//printing membership function values
-			for (i = 0; i < m_functions.size(); i++) {
-				out<<m_functions[i]->evaluate(index)<<"\t";
-			}
-
-
-			out<<endl;
-		}
-
-	}
+  
+  if (!m_functions.empty()) {
+    float min = m_functions[0]->get_lower_bound(); 
+    float max = m_functions[0]->get_upper_bound();
+    int i;
+    
+    //finding limits
+    for (i = 0; i < m_functions.size(); i++) {
+      
+      if (min > m_functions[i]->get_lower_bound()) {
+	min = m_functions[i]->get_lower_bound();
+      }
+      
+      if (max < m_functions[i]->get_upper_bound()) {
+	max = m_functions[i]->get_upper_bound();
+      }
+      
+      
+    }
+    
+    //printing membership values
+    for (float index = min; index <= max; index++) {
+      
+      //printing membership function names
+      for (i = 0; i < m_functions.size(); i++) {
+	out<<m_functions[i]->get_name()<<"\t";	
+      }
+      
+      out<<index<<"\t";
+      
+      
+      //printing membership function values
+      for (i = 0; i < m_functions.size(); i++) {
+	out<<m_functions[i]->evaluate(index)<<"\t";
+      }
+      
+      
+      out<<endl;
+    }
+    
+  }
 }
